Fixed: Respect Quality cutoff if Custom Format cutoff isn't met
Closes #7132
