+++ src/NzbDrone.Core.Test/DecisionEngineTests/UpgradeDiskSpecificationFixture.cs
@@ -4,10 +4,12 @@ using FizzWare.NBuilder;
 using FluentAssertions;
 using Moq;
 using NUnit.Framework;
+using NzbDrone.Common.Serializer;
 using NzbDrone.Core.CustomFormats;
 using NzbDrone.Core.DecisionEngine.Specifications;
 using NzbDrone.Core.Languages;
 using NzbDrone.Core.MediaFiles;
+using NzbDrone.Core.Parser;
 using NzbDrone.Core.Parser.Model;
 using NzbDrone.Core.Profiles.Qualities;
 using NzbDrone.Core.Qualities;
@@ -72,6 +74,42 @@ namespace NzbDrone.Core.Test.DecisionEngineTests
             Mocker.GetMock<ICustomFormatCalculationService>()
                   .Setup(x => x.ParseCustomFormat(It.IsAny<EpisodeFile>()))
                   .Returns(new List<CustomFormat>());
+        }
+
+        private void GivenProfile(QualityProfile profile)
+        {
+            CustomFormatsTestHelpers.GivenCustomFormats();
+            profile.FormatItems = CustomFormatsTestHelpers.GetSampleFormatItems();
+            profile.MinFormatScore = 0;
+            _parseResultMulti.Series.QualityProfile = profile;
+            _parseResultSingle.Series.QualityProfile = profile;
+
+            Console.WriteLine(profile.ToJson());
+        }
+
+        private void GivenFileQuality(QualityModel quality)
+        {
+            _firstFile.Quality = quality;
+            _secondFile.Quality = quality;
+        }
+
+        private void GivenNewQuality(QualityModel quality)
+        {
+            _parseResultMulti.ParsedEpisodeInfo.Quality = quality;
+            _parseResultSingle.ParsedEpisodeInfo.Quality = quality;
+        }
+
+        private void GivenOldCustomFormats(List<CustomFormat> formats)
+        {
+            Mocker.GetMock<ICustomFormatCalculationService>()
+                .Setup(x => x.ParseCustomFormat(It.IsAny<EpisodeFile>()))
+                .Returns(formats);
+        }
+
+        private void GivenNewCustomFormats(List<CustomFormat> formats)
+        {
+            _parseResultMulti.CustomFormats = formats;
+            _parseResultSingle.CustomFormats = formats;
         }
 
         private void WithFirstFileUpgradable()
@@ -154,6 +192,178 @@ namespace NzbDrone.Core.Test.DecisionEngineTests
             _firstFile.Quality = new QualityModel(Quality.WEBDL1080p, new Revision(2));
             _parseResultSingle.ParsedEpisodeInfo.Quality = new QualityModel(Quality.WEBDL1080p);
             _upgradeDisk.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeFalse();
+        }
+
+        [Test]
+        public void should_return_false_if_current_episode_is_equal_to_cutoff()
+        {
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.HDTV720p.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                UpgradeAllowed = true
+            });
+
+            GivenFileQuality(new QualityModel(Quality.HDTV720p, new Revision(version: 2)));
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeFalse();
+        }
+
+        [Test]
+        public void should_return_false_if_current_episode_is_greater_than_cutoff()
+        {
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.HDTV720p.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                UpgradeAllowed = true
+            });
+
+            GivenFileQuality(new QualityModel(Quality.Bluray1080p, new Revision(version: 2)));
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeFalse();
+        }
+
+        [Test]
+        public void should_return_true_when_new_episode_is_proper_but_existing_is_not()
+        {
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.HDTV720p.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                UpgradeAllowed = true
+            });
+
+            GivenFileQuality(new QualityModel(Quality.HDTV720p, new Revision(version: 1)));
+            GivenNewQuality(new QualityModel(Quality.HDTV720p, new Revision(version: 2)));
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeTrue();
+        }
+
+        [Test]
+        public void should_return_false_if_cutoff_is_met_and_quality_is_higher()
+        {
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.HDTV720p.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                UpgradeAllowed = true
+            });
+
+            GivenFileQuality(new QualityModel(Quality.HDTV720p, new Revision(version: 2)));
+            GivenNewQuality(new QualityModel(Quality.Bluray1080p, new Revision(version: 2)));
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeFalse();
+        }
+
+        [Test]
+        public void should_return_false_if_quality_cutoff_is_met_and_quality_is_higher_but_language_is_met()
+        {
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.HDTV720p.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                UpgradeAllowed = true
+            });
+
+            GivenFileQuality(new QualityModel(Quality.HDTV720p, new Revision(version: 2)));
+            GivenNewQuality(new QualityModel(Quality.Bluray1080p, new Revision(version: 2)));
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeFalse();
+        }
+
+        [Test]
+        public void should_return_false_if_cutoff_is_met_and_quality_is_higher_and_language_is_higher()
+        {
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.HDTV720p.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                UpgradeAllowed = true
+            });
+
+            GivenFileQuality(new QualityModel(Quality.HDTV720p, new Revision(version: 2)));
+            GivenNewQuality(new QualityModel(Quality.Bluray1080p, new Revision(version: 2)));
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeFalse();
+        }
+
+        [Test]
+        public void should_return_true_if_cutoff_is_not_met_and_new_quality_is_higher_and_language_is_higher()
+        {
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.HDTV720p.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                UpgradeAllowed = true
+            });
+
+            GivenFileQuality(new QualityModel(Quality.SDTV, new Revision(version: 2)));
+            GivenNewQuality(new QualityModel(Quality.Bluray1080p, new Revision(version: 2)));
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeTrue();
+        }
+
+        [Test]
+        public void should_return_true_if_cutoff_is_not_met_and_language_is_higher()
+        {
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.HDTV720p.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                UpgradeAllowed = true
+            });
+
+            GivenFileQuality(new QualityModel(Quality.SDTV, new Revision(version: 2)));
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeTrue();
+        }
+
+        [Test]
+        public void should_return_false_if_custom_formats_is_met_and_quality_and_format_higher()
+        {
+            var customFormat = new CustomFormat("My Format", new ResolutionSpecification { Value = (int)Resolution.R1080p }) { Id = 1 };
+
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.HDTV720p.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                MinFormatScore = 0,
+                FormatItems = CustomFormatsTestHelpers.GetSampleFormatItems("My Format"),
+                UpgradeAllowed = true
+            });
+
+            GivenFileQuality(new QualityModel(Quality.HDTV720p));
+            GivenNewQuality(new QualityModel(Quality.Bluray1080p));
+
+            GivenOldCustomFormats(new List<CustomFormat>());
+            GivenNewCustomFormats(new List<CustomFormat> { customFormat });
+
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeFalse();
+        }
+
+        [Test]
+        public void should_return_true_if_cutoffs_are_met_but_is_a_revision_upgrade()
+        {
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.HDTV1080p.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                UpgradeAllowed = true
+            });
+
+            GivenFileQuality(new QualityModel(Quality.WEBDL1080p, new Revision(version: 1)));
+            GivenNewQuality(new QualityModel(Quality.WEBDL1080p, new Revision(version: 2)));
+
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeTrue();
+        }
+
+        [Test]
+        public void should_return_false_if_quality_profile_does_not_allow_upgrades_but_cutoff_is_set_to_highest_quality()
+        {
+            GivenProfile(new QualityProfile
+            {
+                Cutoff = Quality.RAWHD.Id,
+                Items = Qualities.QualityFixture.GetDefaultQualities(),
+                UpgradeAllowed = false
+            });
+
+            GivenFileQuality(new QualityModel(Quality.WEBDL1080p));
+            GivenNewQuality(new QualityModel(Quality.Bluray1080p));
+
+            Subject.IsSatisfiedBy(_parseResultSingle, null).Accepted.Should().BeFalse();
         }
     }
 }
