--- src/NzbDrone.Core/DecisionEngine/Specifications/CutoffSpecification.cs
@@ -1,58 +1,0 @@
-using System.Linq;
-using NLog;
-using NzbDrone.Core.CustomFormats;
-using NzbDrone.Core.IndexerSearch.Definitions;
-using NzbDrone.Core.Parser.Model;
-
-namespace NzbDrone.Core.DecisionEngine.Specifications
-{
-    public class CutoffSpecification : IDecisionEngineSpecification
-    {
-        private readonly UpgradableSpecification _upgradableSpecification;
-        private readonly ICustomFormatCalculationService _formatService;
-        private readonly Logger _logger;
-
-        public CutoffSpecification(UpgradableSpecification upgradableSpecification, ICustomFormatCalculationService formatService, Logger logger)
-        {
-            _upgradableSpecification = upgradableSpecification;
-            _formatService = formatService;
-            _logger = logger;
-        }
-
-        public SpecificationPriority Priority => SpecificationPriority.Default;
-        public RejectionType Type => RejectionType.Permanent;
-
-        public virtual Decision IsSatisfiedBy(RemoteEpisode subject, SearchCriteriaBase searchCriteria)
-        {
-            var qualityProfile = subject.Series.QualityProfile.Value;
-
-            foreach (var file in subject.Episodes.Where(c => c.EpisodeFileId != 0).Select(c => c.EpisodeFile.Value))
-            {
-                if (file == null)
-                {
-                    _logger.Debug("File is no longer available, skipping this file.");
-                    continue;
-                }
-
-                _logger.Debug("Comparing file quality with report. Existing file is {0}", file.Quality);
-
-                var customFormats = _formatService.ParseCustomFormat(file);
-
-                if (!_upgradableSpecification.CutoffNotMet(qualityProfile,
-                                                           file.Quality,
-                                                           _formatService.ParseCustomFormat(file),
-                                                           subject.ParsedEpisodeInfo.Quality))
-                {
-                    _logger.Debug("Cutoff already met, rejecting.");
-
-                    var qualityCutoffIndex = qualityProfile.GetIndex(qualityProfile.Cutoff);
-                    var qualityCutoff = qualityProfile.Items[qualityCutoffIndex.Index];
-
-                    return Decision.Reject("Existing file meets cutoff: {0}", qualityCutoff);
-                }
-            }
-
-            return Decision.Accept();
-        }
-    }
-}
